requirements.txt
flask
python-dotenv
elasticsearch
elasticsearch_dsl